uint32_t standardisation found array boundary errors resized Wrms arrays

+++ src/ISR.cpp
@@ -1,5 +1,5 @@
 #include "extern.h"
-long unsigned int waitStart, appStart;
+uint32_t waitStart, appStart;
 
  
 void ISRwatchDog () {
